+++ src/modules/students/views/StudentsMaterial/components/students-recordings-list.vue
@@ -1,0 +1,18 @@
+<template>
+    <StudentsMaterialsList />
+</template>
+  <script>
+  
+  export default {
+    name: 'StudentsMaterialsList',
+    components: {
+      StudentsMaterialsList: () =>
+        import(/* webpackChunkName: "StudentsMaterialsList" */ './students-materials-base.vue')
+    },
+    data() {
+      return {
+        reloadDatatableUsers: false
+      }
+    }
+  }
+  </script>
