+++ src/modules/students/views/StudentsMaterial/students-recordings.store.js
@@ -1,0 +1,33 @@
+import DataTableStore from '@/modules/resources/store/data-table.store'
+
+export default {
+  name: 'studentsRecordingsStore',
+  namespaced: true,
+  state: {
+    ...DataTableStore.state,
+    currentUserForUpdate: null,
+    tags: [],
+    lessons: []
+  },
+  getters: {
+    ...DataTableStore.getters
+  },
+
+  mutations: {
+    ...DataTableStore.mutations,
+    SET_TAGS(state, payload) {
+      state.tags = payload
+    },
+    SET_LESSONS(state, payload) {
+      state.lessons = payload
+    }
+  },
+  actions: {
+    ...DataTableStore.actions,
+    resetTableOptions({ commit }) {
+      DataTableStore.actions.resetTableOptions({ commit })
+      commit('SET_TAGS', [])
+      commit('SET_LESSONS', [])
+    }
+  }
+}
